LayoutTests/media/video-load-networkState.html

<video controls></video>
<script src=video-test.js></script>
<script>
test("video.networkState == HTMLMediaElement.EMPTY");
waitForEventAndTest("loadstart", "video.networkState == HTMLMediaElement.LOADING");
waitForEventAndTest("loadedmetadata", "video.networkState == HTMLMediaElement.LOADED_METADATA");
waitForEventAndTest("loadedfirstframe", "video.networkState == HTMLMediaElement.LOADED_FIRST_FRAME");
waitForEventTestAndEnd("load", "video.networkState == HTMLMediaElement.LOADED");
video.src = "content/test.mp4";
</script>
